Skill validation moved into the clientValidation file
