+++ src/App.module.css
@@ -193,8 +193,13 @@ input {
   color: #0d5890;
 }
 
-.newSkillLink {
-  list-style: none;
+.newSkillButton {
+  background-color: rgb(0, 0, 0);
+  margin: 1%;
+  padding: 1%;
+  border-radius: 10px;
+  border-color: #0d5890;
+  color:white;
 }
