+++ src/clientValidation.js
@@ -28,8 +28,27 @@ function checkDisplayName(displayName){
     return displayName
 }
 
+function checkSkillname(skillname){
+    if(!skillname) throw "No skill name provided"
+    if(typeof skillname!=='string') throw "Skill name must be a string"
+    skillname=xss(skillname.trim())
+    if(skillname.length<3) throw "Skill name must be at least 3 characters long"
+
+    return skillname
+}
+
+function checkSkilllevel(skilllevel){
+    if(!skilllevel) throw "No skill level provided"
+    if(typeof skilllevel!=='string') throw "Skill level must be a string"
+    skilllevel=xss(skilllevel.trim())
+
+    return skilllevel
+}
+
 export default {
     checkUsername,
     checkPassword,
-    checkDisplayName
+    checkDisplayName,
+    checkSkillname,
+    checkSkilllevel
 }
